requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
jupyter
python-dotenv
pyyaml
statsmodels
xgboost
